feat(src/*): refactor folders structure to add support to ts and js plane files
