+++ src/index.ts
@@ -1,4 +1,6 @@
+import { jsConfig as jsEslint } from './js/index.config.ts'
 import { reactConfig as reactEslint } from './react/index.config.ts'
 import { reactTsConfig as reactTsEslint } from './react-ts/index.config.ts'
+import { tsConfig as tsEslint } from './ts/index.config.ts'
 
-export { reactEslint, reactTsEslint }
+export { jsEslint, tsEslint, reactEslint, reactTsEslint }
