+++ src/js/eslint.config.js
@@ -1,0 +1,4 @@
+// @ts-check
+import { tsEslint } from './../../dist/esm/index.mjs'
+
+export default tsEslint
