+++ src/react-ts/eslint.config.js
@@ -1,0 +1,4 @@
+// @ts-check
+import { reactTsEslint } from './../../dist/esm/index.mjs'
+
+export default reactTsEslint
