+++ src/react/eslint.config.js
@@ -1,0 +1,4 @@
+// @ts-check
+import { reactEslint } from './../../dist/esm/index.mjs'
+
+export default reactEslint
